job_profiling/requirements.txt
django==3.1.2
pymongo==3.11.2
mongoengine==0.23.1
django-mongoengine==0.4.6
jango-rest-framework-mongoengine==3.4.1
